// src/modules/iE/reports/style-change-over/style-change-over-type.ts
export interface IStyleChangeOver {
  ENTRY_DATE: Date;
  ENTRY_NO: string;
  LAYOUT_START_TIME: Date;
  LAYOUT_END_TIME: Date;
  BUYER_ID: number;
  STYLE_ID: number;
  PO_ID: number;
  FLOOR_ID: number;
  LINE_ID: number;
  PRODUCTION_TYPE: string;
  SMV: string;
  TECHNICIAN_ID: number;
  SUPERVISOR_ID: number;
  IE_ID: number;
  OPERATOR: string;
  HP: string;
  IRON_MAN: string;
  CREATED_BY: string;
  UPDATED_BY: string;
  ID: number;
  BUYER_NAME: string;
  STYLE_NO: string;
  PO_NO: string;
  FLOOR_NAME: string;
  LINE_NAME: string;
  TECHNICIAN_NAME: string;
  SUPERVISOR_NAME: string;
  IE_NAME: string;
  REASON: string;
  TOTAL_TIME: string;
  REMARKS: string;
  COMPANY_NAME: string;
  COMPANY_ADDRESS: string;
  REQ_OP: number;
  REQ_HP: number;
  REQ_IR: number;
  ITEM_TYPE: string;
  ITEM: string;
}